Delete in frontend uses JWT

+++ src/App.tsx
@@ -61,16 +61,25 @@ const App = () => {
 		setUsers(userResponse)
 	}
 	const handleDeleteUser = async (userId: string): Promise<void> => {
+		const jwt: string | null = localStorage.getItem(LS_KEY)
+		if( !jwt ) {
+			console.log('No JWT in localStorage')
+			return
+		}
+		
 		const response: Response = await fetch('/api/users/' + userId, {
-			method: 'DELETE'
-			// TODO: lägg till JWT här!
+			method: 'DELETE',
+			headers: {
+				'Authorization': `Bearer: ${jwt}`
+			}
 		})
 
 		// kontrollera status för responsen. Lyckades requestet?
 		if( response.status === 204 ) {
 			console.log('DELETE lyckades!')
 			handleGetUsers() // uppdatera listan
-			// En alternativ metod: ta bort användaren direkt ur state-variabeln
+			// En alternativ metod: ta bort användaren direkt ur state-variabeln <- går inte, eftersom vi inte har userId
+			// Tips! Låt servern skicka tillbaka en ny JWT med varje request - så kan vi uppdatera JWT i localStorage, så att det dröjer längre tills JWT slutar gälla (återställa timeout)
 
 		} else {
 			console.log('DELETE failade med status ' + response.status)
